Update application.yaml
Edit the chrome cleanup, now the chrome cleanup will clear everything but will not reset as doing before. In the previous version chrome cleanup was removing everything as the user was seeing chrome first welcome and all the stuff when we install chrome at the first time, now will clear everything useless and keep the files to chrome read and avoid the first welcome screen.